+++ clients/horizonclient/trade_request.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"net/http"
 	"net/url"
 
 	hProtocol "github.com/stellar/go/protocols/horizon"
@@ -60,6 +61,16 @@ func (tr TradeRequest) BuildURL() (endpoint string, err error) {
 	}
 
 	return endpoint, err
+}
+
+// HTTPRequest returns the http request for the trades endpoint
+func (tr TradeRequest) HTTPRequest(horizonURL string) (*http.Request, error) {
+	endpoint, err := tr.BuildURL()
+	if err != nil {
+		return nil, err
+	}
+
+	return http.NewRequest("GET", horizonURL+endpoint, nil)
 }
 
 // TradeHandler is a function that is called when a new trade is received