+++ services/regulated-assets-approval-server/internal/serve/tx_approve.go
@@ -59,15 +59,6 @@ func (h txApproveHandler) validate() error {
 	return nil
 }
 
-func convertThresholdToReadableString(threshold int64) (string, error) {
-	thresholdStr := amount.StringFromInt64(threshold)
-	res, err := strconv.ParseFloat(thresholdStr, 1)
-	if err != nil {
-		return "", errors.Wrap(err, "converting threshold amount from string to float")
-	}
-	return fmt.Sprintf("%.2f", res), nil
-}
-
 func (h txApproveHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	err := h.validate()
@@ -95,9 +86,8 @@ func (h txApproveHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	txApproveResp.Render(w)
 }
 
-// validateInput performs some validations on the provided transaction. It can
-// reject the transaction based on general criteria that would be applied in any
-// approval server.
+// validateInput validates if the input parameters contain a valid transaction
+// and if the source account is not set in a way that would harm the issuer.
 func (h txApproveHandler) validateInput(ctx context.Context, in txApproveRequest) (*txApprovalResponse, *txnbuild.Transaction) {
 	if in.Tx == "" {
 		log.Ctx(ctx).Error(`request is missing parameter "tx".`)
@@ -117,19 +107,20 @@ func (h txApproveHandler) validateInput(ctx context.Context, in txApproveRequest
 	}
 
 	if tx.SourceAccount().AccountID == h.issuerKP.Address() {
-		log.Ctx(ctx).Errorf("transaction %s sourceAccount is the same as the server issuer account %s",
-			in.Tx,
-			h.issuerKP.Address())
-		return NewRejectedTxApprovalResponse("The source account is invalid."), nil
+		log.Ctx(ctx).Errorf("transaction sourceAccount is the same as the server issuer account %s", h.issuerKP.Address())
+		return NewRejectedTxApprovalResponse("Transaction source account is invalid."), nil
 	}
 
-	if len(tx.Operations()) != 1 {
-		return NewRejectedTxApprovalResponse("Please submit a transaction with exactly one operation of type payment."), nil
-	}
+	// only AllowTrust operations can have the issuer as their source account
+	for _, op := range tx.Operations() {
+		if _, ok := op.(*txnbuild.AllowTrust); ok {
+			continue
+		}
 
-	if tx.Operations()[0].GetSourceAccount() == h.issuerKP.Address() {
-		log.Ctx(ctx).Error(`transaction contains one or more operations where sourceAccount is issuer account.`)
-		return NewRejectedTxApprovalResponse("There is one or more unauthorized operations in the provided transaction."), nil
+		if op.GetSourceAccount() == h.issuerKP.Address() {
+			log.Ctx(ctx).Error("transaction contains one or more unauthorized operations where source account is the issuer account")
+			return NewRejectedTxApprovalResponse("There are one or more unauthorized operations in the provided transaction."), nil
+		}
 	}
 
 	return nil, tx
@@ -145,14 +136,27 @@ func (h txApproveHandler) txApprove(ctx context.Context, in txApproveRequest) (r
 		log.Ctx(ctx).Debug("====  did log responses ====")
 	}()
 
-	txRejectedResp, tx := h.validateInput(ctx, in)
-	if txRejectedResp != nil {
-		return txRejectedResp, nil
+	rejectedResponse, tx := h.validateInput(ctx, in)
+	if rejectedResponse != nil {
+		return rejectedResponse, nil
+	}
+
+	txSuccessResp, err := h.handleSuccessResponseIfNeeded(ctx, tx)
+	if err != nil {
+		return nil, errors.Wrap(err, "checking if transaction in request was compliant")
+	}
+	if txSuccessResp != nil {
+		return txSuccessResp, nil
+	}
+
+	// validate the revisable transaction has one operation.
+	if len(tx.Operations()) != 1 {
+		return NewRejectedTxApprovalResponse("Please submit a transaction with exactly one operation of type payment."), nil
 	}
 
 	paymentOp, ok := tx.Operations()[0].(*txnbuild.Payment)
 	if !ok {
-		log.Ctx(ctx).Error(`transaction contains one or more operations is not of type payment`)
+		log.Ctx(ctx).Error("transaction does not contain a payment operation")
 		return NewRejectedTxApprovalResponse("There is one or more unauthorized operations in the provided transaction."), nil
 	}
 	paymentSource := paymentOp.SourceAccount
@@ -160,6 +164,11 @@ func (h txApproveHandler) txApprove(ctx context.Context, in txApproveRequest) (r
 		paymentSource = tx.SourceAccount().AccountID
 	}
 
+	if paymentOp.Destination == h.issuerKP.Address() {
+		return NewRejectedTxApprovalResponse("Can't transfer asset to its issuer."), nil
+	}
+
+	// validate payment asset is the one supported by the issuer
 	issuerAddress := h.issuerKP.Address()
 	if paymentOp.Asset.GetCode() != h.assetCode || paymentOp.Asset.GetIssuer() != issuerAddress {
 		log.Ctx(ctx).Error(`the payment asset is not supported by this issuer`)
@@ -168,8 +177,9 @@ func (h txApproveHandler) txApprove(ctx context.Context, in txApproveRequest) (r
 
 	acc, err := h.horizonClient.AccountDetail(horizonclient.AccountRequest{AccountID: paymentSource})
 	if err != nil {
-		return nil, errors.Wrapf(err, "getting detail for payment source account %s", issuerAddress)
+		return nil, errors.Wrapf(err, "getting detail for payment source account %s", paymentSource)
 	}
+
 	// validate the sequence number
 	accountSequence, err := strconv.ParseInt(acc.Sequence, 10, 64)
 	if err != nil {
@@ -179,15 +189,15 @@ func (h txApproveHandler) txApprove(ctx context.Context, in txApproveRequest) (r
 		log.Ctx(ctx).Errorf(`invalid transaction sequence number tx.SourceAccount().Sequence: %d, accountSequence+1:%d`, tx.SourceAccount().Sequence, accountSequence+1)
 		return NewRejectedTxApprovalResponse("Invalid transaction sequence number."), nil
 	}
-	// Validate if payment operation requires KYC.
-	var kycRequiredResponse *txApprovalResponse
-	kycRequiredResponse, err = h.handleKYCRequiredOperationIfNeeded(ctx, paymentSource, paymentOp)
+
+	actionRequiredResponse, err := h.handleActionRequiredResponseIfNeeded(ctx, paymentSource, paymentOp)
 	if err != nil {
 		return nil, errors.Wrap(err, "handling KYC required payment")
 	}
-	if kycRequiredResponse != nil {
-		return kycRequiredResponse, nil
+	if actionRequiredResponse != nil {
+		return actionRequiredResponse, nil
 	}
+
 	// build the transaction
 	revisedOperations := []txnbuild.Operation{
 		&txnbuild.AllowTrust{
@@ -231,7 +241,6 @@ func (h txApproveHandler) txApprove(ctx context.Context, in txApproveRequest) (r
 	if err != nil {
 		return nil, errors.Wrap(err, "signing transaction")
 	}
-
 	txe, err := revisedTx.Base64()
 	if err != nil {
 		return nil, errors.Wrap(err, "encoding revised transaction")
@@ -240,14 +249,14 @@ func (h txApproveHandler) txApprove(ctx context.Context, in txApproveRequest) (r
 	return NewRevisedTxApprovalResponse(txe), nil
 }
 
-// handleKYCRequiredOperationIfNeeded validates and returns an action_required response if the payment requires KYC.
-func (h txApproveHandler) handleKYCRequiredOperationIfNeeded(ctx context.Context, stellarAddress string, paymentOp *txnbuild.Payment) (*txApprovalResponse, error) {
-	// validate payment operation against KYC condition(s).
-	KYCRequiredMessage, err := h.kycRequiredMessageIfNeeded(paymentOp)
+// handleActionRequiredResponseIfNeeded validates and returns an action_required
+// response if the payment requires KYC.
+func (h txApproveHandler) handleActionRequiredResponseIfNeeded(ctx context.Context, stellarAddress string, paymentOp *txnbuild.Payment) (*txApprovalResponse, error) {
+	paymentAmount, err := amount.ParseInt64(paymentOp.Amount)
 	if err != nil {
-		return nil, errors.Wrap(err, "validating KYC")
+		return nil, errors.Wrap(err, "parsing payment amount from string to Int64")
 	}
-	if KYCRequiredMessage == "" {
+	if paymentAmount <= h.kycThreshold {
 		return nil, nil
 	}
 
@@ -259,17 +268,17 @@ func (h txApproveHandler) handleKYCRequiredOperationIfNeeded(ctx context.Context
 			ON CONFLICT(stellar_address) DO NOTHING
 			RETURNING *
 		)
-		SELECT callback_id, approved_at, rejected_at FROM new_row
+		SELECT callback_id, approved_at, rejected_at, pending_at FROM new_row
 		UNION
-		SELECT callback_id, approved_at, rejected_at
+		SELECT callback_id, approved_at, rejected_at, pending_at
 		FROM accounts_kyc_status
 		WHERE stellar_address = $1
 	`
 	var (
-		callbackID             string
-		approvedAt, rejectedAt sql.NullTime
+		callbackID                        string
+		approvedAt, rejectedAt, pendingAt sql.NullTime
 	)
-	err = h.db.QueryRowContext(ctx, q, stellarAddress, intendedCallbackID).Scan(&callbackID, &approvedAt, &rejectedAt)
+	err = h.db.QueryRowContext(ctx, q, stellarAddress, intendedCallbackID).Scan(&callbackID, &approvedAt, &rejectedAt, &pendingAt)
 	if err != nil {
 		return nil, errors.Wrap(err, "inserting new row into accounts_kyc_status table")
 	}
@@ -277,34 +286,154 @@ func (h txApproveHandler) handleKYCRequiredOperationIfNeeded(ctx context.Context
 	if approvedAt.Valid {
 		return nil, nil
 	}
+
+	kycThreshold, err := convertAmountToReadableString(h.kycThreshold)
+	if err != nil {
+		return nil, errors.Wrap(err, "converting kycThreshold to human readable string")
+	}
+
 	if rejectedAt.Valid {
-		kycThreshold, err := convertThresholdToReadableString(h.kycThreshold)
-		if err != nil {
-			return nil, errors.Wrap(err, "converting kycThreshold to human readable string")
-		}
 		return NewRejectedTxApprovalResponse(fmt.Sprintf("Your KYC was rejected and you're not authorized for operations above %s %s.", kycThreshold, h.assetCode)), nil
 	}
 
+	if pendingAt.Valid {
+		return NewPendingTxApprovalResponse(fmt.Sprintf("Your account could not be verified as approved nor rejected and was marked as pending. You will need staff authorization for operations above %s %s.", kycThreshold, h.assetCode)), nil
+	}
+
 	return NewActionRequiredTxApprovalResponse(
-		KYCRequiredMessage,
+		fmt.Sprintf(`Payments exceeding %s %s require KYC approval. Please provide an email address.`, kycThreshold, h.assetCode),
 		fmt.Sprintf("%s/kyc-status/%s", h.baseURL, callbackID),
 		[]string{"email_address"},
 	), nil
 }
 
-// kycRequiredMessageIfNeeded returns a "action_required" message for the NewActionRequiredTxApprovalResponse if the payment operation meets KYC conditions.
-// Currently rule(s) are, checking if payment amount is > KYCThreshold amount.
-func (h txApproveHandler) kycRequiredMessageIfNeeded(paymentOp *txnbuild.Payment) (string, error) {
-	paymentAmount, err := amount.ParseInt64(paymentOp.Amount)
+// handleSuccessResponseIfNeeded inspects the incoming transaction and returns a
+// "success" response if it's already compliant with the SEP-8 authorization spec.
+func (h txApproveHandler) handleSuccessResponseIfNeeded(ctx context.Context, tx *txnbuild.Transaction) (*txApprovalResponse, error) {
+	if len(tx.Operations()) != 5 {
+		return nil, nil
+	}
+
+	rejectedResp, paymentOp, paymentSource := validateTransactionOperationsForSuccess(ctx, tx, h.issuerKP.Address())
+	if rejectedResp != nil {
+		return rejectedResp, nil
+	}
+
+	if paymentOp.Destination == h.issuerKP.Address() {
+		return NewRejectedTxApprovalResponse("Can't transfer asset to its issuer."), nil
+	}
+
+	// pull current account details from the network then validate the tx sequence number
+	acc, err := h.horizonClient.AccountDetail(horizonclient.AccountRequest{AccountID: paymentSource})
 	if err != nil {
-		return "", errors.Wrap(err, "parsing account payment amount from string to Int64")
+		return nil, errors.Wrapf(err, "getting detail for payment source account %s", paymentSource)
 	}
-	if paymentAmount > h.kycThreshold {
-		kycThreshold, err := convertThresholdToReadableString(h.kycThreshold)
-		if err != nil {
-			return "", errors.Wrap(err, "converting kycThreshold to human readable string")
+	accountSequence, err := strconv.ParseInt(acc.Sequence, 10, 64)
+	if err != nil {
+		return nil, errors.Wrapf(err, "parsing account sequence number %q from string to int64", acc.Sequence)
+	}
+	if tx.SourceAccount().Sequence != accountSequence+1 {
+		log.Ctx(ctx).Errorf(`invalid transaction sequence number tx.SourceAccount().Sequence: %d, accountSequence+1: %d`, tx.SourceAccount().Sequence, accountSequence+1)
+		return NewRejectedTxApprovalResponse("Invalid transaction sequence number."), nil
+	}
+
+	kycRequiredResponse, err := h.handleActionRequiredResponseIfNeeded(ctx, paymentSource, paymentOp)
+	if err != nil {
+		return nil, errors.Wrap(err, "handling KYC required payment")
+	}
+	if kycRequiredResponse != nil {
+		return kycRequiredResponse, nil
+	}
+
+	// sign transaction with issuer's signature and encode it
+	tx, err = tx.Sign(h.networkPassphrase, h.issuerKP)
+	if err != nil {
+		return nil, errors.Wrap(err, "signing transaction")
+	}
+	txe, err := tx.Base64()
+	if err != nil {
+		return nil, errors.Wrap(err, "encoding revised transaction")
+	}
+
+	return NewSuccessTxApprovalResponse(txe, "Transaction is compliant and signed by the issuer."), nil
+}
+
+// validateTransactionOperationsForSuccess checks if the incoming transaction
+// operations are compliant with the anchor's SEP-8 policy.
+func validateTransactionOperationsForSuccess(ctx context.Context, tx *txnbuild.Transaction, issuerAddress string) (resp *txApprovalResponse, paymentOp *txnbuild.Payment, paymentSource string) {
+	if len(tx.Operations()) != 5 {
+		return NewRejectedTxApprovalResponse("Unsupported number of operations."), nil, ""
+	}
+
+	// extract the payment operation and payment source account.
+	paymentOp, ok := tx.Operations()[2].(*txnbuild.Payment)
+	if !ok {
+		log.Ctx(ctx).Error(`third operation is not of type payment`)
+		return NewRejectedTxApprovalResponse("There are one or more unexpected operations in the provided transaction."), nil, ""
+	}
+	paymentSource = paymentOp.SourceAccount
+	if paymentSource == "" {
+		paymentSource = tx.SourceAccount().AccountID
+	}
+
+	assetCode := paymentOp.Asset.GetCode()
+
+	operationsValid := func() bool {
+		op0, ok := tx.Operations()[0].(*txnbuild.AllowTrust)
+		if !ok ||
+			op0.Trustor != paymentSource ||
+			op0.Type.GetCode() != assetCode ||
+			!op0.Authorize ||
+			op0.SourceAccount != issuerAddress {
+			return false
 		}
-		return fmt.Sprintf(`Payments exceeding %s %s requires KYC approval. Please provide an email address.`, kycThreshold, h.assetCode), nil
+
+		op1, ok := tx.Operations()[1].(*txnbuild.AllowTrust)
+		if !ok ||
+			op1.Trustor != paymentOp.Destination ||
+			op1.Type.GetCode() != assetCode ||
+			!op1.Authorize ||
+			op1.SourceAccount != issuerAddress {
+			return false
+		}
+
+		op2, ok := tx.Operations()[2].(*txnbuild.Payment)
+		if !ok || op2 != paymentOp {
+			return false
+		}
+
+		op3, ok := tx.Operations()[3].(*txnbuild.AllowTrust)
+		if !ok ||
+			op3.Trustor != paymentOp.Destination ||
+			op3.Type.GetCode() != assetCode ||
+			op3.Authorize ||
+			op3.SourceAccount != issuerAddress {
+			return false
+		}
+
+		op4, ok := tx.Operations()[4].(*txnbuild.AllowTrust)
+		if !ok ||
+			op4.Trustor != paymentSource ||
+			op4.Type.GetCode() != assetCode ||
+			op4.Authorize ||
+			op4.SourceAccount != issuerAddress {
+			return false
+		}
+
+		return true
+	}()
+	if !operationsValid {
+		return NewRejectedTxApprovalResponse("There are one or more unexpected operations in the provided transaction."), nil, ""
 	}
-	return "", nil
+
+	return nil, paymentOp, paymentSource
+}
+
+func convertAmountToReadableString(threshold int64) (string, error) {
+	amountStr := amount.StringFromInt64(threshold)
+	amountFloat, err := strconv.ParseFloat(amountStr, 64)
+	if err != nil {
+		return "", errors.Wrap(err, "converting threshold amount from string to float")
+	}
+	return fmt.Sprintf("%.2f", amountFloat), nil
 }