+++ txnbuild/transaction.go
@@ -76,6 +76,13 @@ func concatSignatures(
 	return extended, nil
 }
 
+func concatSignatureDecorated(e xdr.TransactionEnvelope, signatures []xdr.DecoratedSignature, newSignatures []xdr.DecoratedSignature) ([]xdr.DecoratedSignature, error) {
+	extended := make([]xdr.DecoratedSignature, len(signatures)+len(newSignatures))
+	copy(extended, signatures)
+	copy(extended[len(signatures):], newSignatures)
+	return extended, nil
+}
+
 func concatSignatureBase64(e xdr.TransactionEnvelope, signatures []xdr.DecoratedSignature, networkStr, publicKey, signature string) ([]xdr.DecoratedSignature, error) {
 	if signature == "" {
 		return nil, errors.New("signature not presented")
@@ -305,6 +312,17 @@ func (t *Transaction) SignHashX(preimage []byte) (*Transaction, error) {
 	return t.clone(extendedSignatures), nil
 }
 
+// AddSignatureDecorated returns a new Transaction instance which extends the current instance
+// with an additional decorated signature(s).
+func (t *Transaction) AddSignatureDecorated(signature ...xdr.DecoratedSignature) (*Transaction, error) {
+	extendedSignatures, err := concatSignatureDecorated(t.envelope, t.Signatures(), signature)
+	if err != nil {
+		return nil, err
+	}
+
+	return t.clone(extendedSignatures), nil
+}
+
 // AddSignatureBase64 returns a new Transaction instance which extends the current instance
 // with an additional signature derived from the given base64-encoded signature.
 func (t *Transaction) AddSignatureBase64(network, publicKey, signature string) (*Transaction, error) {
@@ -346,10 +364,14 @@ func (t *Transaction) ClaimableBalanceID(operationIndex int) (string, error) {
 
 	// We mimic the relevant code from Stellar Core
 	// https://github.com/stellar/stellar-core/blob/9f3cc04e6ec02c38974c42545a86cdc79809252b/src/test/TestAccount.cpp#L285
+	//
+	// Note that the source account must be *unmuxed* for this to work.
+	muxedAccountId := xdr.MustMuxedAddress(t.sourceAccount.AccountID).ToAccountId()
+	gAddress := muxedAccountId.Address()
 	operationId := xdr.OperationId{
 		Type: xdr.EnvelopeTypeEnvelopeTypeOpId,
 		Id: &xdr.OperationIdId{
-			SourceAccount: xdr.MustMuxedAddress(t.sourceAccount.AccountID),
+			SourceAccount: xdr.MustMuxedAddress(gAddress),
 			SeqNum:        xdr.SequenceNumber(t.sourceAccount.Sequence),
 			OpNum:         xdr.Uint32(operationIndex),
 		},